Basic Functionality Setup
I implemented basic functionalities for displaying and creating posts in the application. This included configuring both the server side and client side. On the server side, I set up an Express.js server and connected it to MongoDB using Mongoose. On the client side, I established pages that serve as interfaces for creating and viewing blog post data.